requirements.txt
django
djangorestframework
djangorestframework_simplejwt
psycopg2-binary
dj-config-url
python-decouple
drf-yasg
django-cors-headers
pillow
django-filter
celery[redis]
loguru
coverage
model_bakery
django-cors-headers
pyngrok
beautifulsoup4
lxml
selenium
webdriver_manager
fake-useragent
geopy
social-auth-app-django
channels
channels-redis
django-allauth
pandas
matplotlib
seaborn
numpy
plotly
absl-py
astor
gast
google-pasta
grpcio
gunicorn
h5py
joblib
Keras
Keras-Applications
Keras-Preprocessing
PyYAML
sklearn
tensorboard
tensorflow
prophet
bokeh
cloudinary
opencv-python